# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Portfolio tracking API",
        "title": "My Portfolio API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/distributions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "distributions"
                ],
                "summary": "List income distributions (dividends/interest) with pagination and a year summary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter to the tax year the income was paid",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to a single holding",
                        "name": "holding_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedDistributions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "distributions"
                ],
                "summary": "Record an income payment by hand",
                "parameters": [
                    {
                        "description": "Distribution payload",
                        "name": "distribution",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DistributionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Distribution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/distributions/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "distributions"
                ],
                "summary": "Delete an income payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Distribution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "distributions"
                ],
                "summary": "Update an income payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Distribution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Distribution payload",
                        "name": "distribution",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DistributionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Distribution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Distribution": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "asset_type": {
                    "type": "string"
                },
                "category": {
                    "description": "Category is the kind of income: \"dividend\" (from a security) or\n\"interest\" (from a cash/savings balance).",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "holding_id": {
                    "description": "HoldingID, Symbol, and AssetType identify the paying security. Symbol and\nAssetType are unset for account-level income like bank interest, and\nHoldingID is nil until a matching holding exists to link to.",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_date": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "transaction_id": {
                    "description": "TransactionID is nil for a cash dividend or interest payment. It's reserved\nfor a reinvested dividend, which also opens a buy lot - not handled yet.",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "DistributionRequest": {
            "type": "object",
            "required": [
                "account_id",
                "payment_date"
            ],
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "asset_type": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "holding_id": {
                    "type": "string"
                },
                "payment_date": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "DistributionSummary": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "number"
                }
            }
        },
...
        "PaginatedDistributions": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Distribution"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "summary": {
                    "$ref": "#/definitions/DistributionSummary"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}